chore: split StackMachine struct into separate file
